Update azure-terraform-command-handler.ts
The use of arm_client_id, arm_client_secret, arm_subscription_id and arm_tenant_id will soon be deprecated from the Terraform Provider in favour of those values without the arm_ in the value.  

This change will remove the warning when running the plugin and also provide future support.