Enable publishing of ExternalTfs to Marketplace (#1314)
The pipeline for publishing extensions on Marketplace
was broken because app registration of
canarytest-extensions-ci-tests-pipeline-connector service connection
was not available (probably deleted from Azure portal).

Error message: ERROR AADSTS53003:
Access has been blocked by Conditional Access policies.
The access policy does not allow token issuance.


As a mitigation step, the old SC is deleted and a new one
with proper app registration is created by using the same name.
The version number of ExternalTfs extension needs to be updated,
because publishing of test extension was successfully done
before pipeline for running CI tests failed and
we won't be able to run it again (version number must be increased).